Repeat fix in #78387 for routes without params
Follow up to #78387. When incremental PPR mode is enabled, we need to
create a special "inverse" route to handle tree requests for every route
in the manifest that doesn't have PPR enabled. #78387 fixed this for
"dynamic" routes — which in the context of the relevant module I
believe means routes with params, and has nothing to do with dynamic
data in the PPR sense — but it didn't handle for "static" routes —
which here means routes without any params.

The fix is to apply the same logic in #78387 for all routes, both
"static" and  "dynamic", in the routes manifest.