[dynamicIO] log dynamic validation errors consistently in dev
In dev when dynamicIO is enabled we performa a validating prerender to look for code that would be a build error in `next build`. However due to the construction of the earlier implementation of this code some validation messages were being suprressed because they were conveyed as a thrown error rather than through console.error. This change updates the mechanism by which we raise these issues in dev so they are consistently messaged.