[dynamicIO] Do not throw when metadata is dynamic
Metadata can be dynamic without that being an error. A bug was introduced in a recent refactor where the presence of dynamic metadata was incorrectly interpretted as a sign that the metadata did not have a Suspense boundary around it. This would be a problem because metadata is now streaming however it was actually a misuse of the tracking flag and would incorrectly error if you had dynamic metadata at all. This fix addresses the incorrect invariant by removing it.