fix: make webpack handle "use cache" in node_modules  (#78606)
Turbopack applies the "use cache" transform in node_modules, but Webpack
wasn't doing that. This is due to a sneaky little regex that controls
whether files in node_modules are passed to SWC for transforming or not
-- `use client` and `use server` were included there, but `use cache`
wasn't. This PR fixes that.

(I don't really expect NPM packages to use `"use cache"` because it's
pretty application-specific, but it might be relevant in a monorepo
situation where shared utils that use the directive can live in a
separate package)

---------

Co-authored-by: Hendrik Liebau <mail@hendrik-liebau.de>