Use React's `prerender` function for `"use cache"` with Dynamic IO (#78382)
During prerendering, we don't need to return a stream early for a `"use
cache"` function. We can instead prerender the cache function fully,
before returning the prelude as the stream. This simplifies the timeout
handling a bit, when `dynamicIO` is enabled, and more importantly, it
prepares us for handling cookies access in `"use cache"` functions
during prerendering.

- [x] depends on facebook/react#32953