Remove useMemo from useSearchParams
I was looking at this hook because we're seeing a lot of unusual errors
in production that originate from here. Not sure what the cause is yet,
so I was trying to look for anything that look fishy.

Some of the stack traces originated from an internal React invariant
inside the useMemo hook. I'm not sure how this is possible, but to
reduce the surface area of things that might go wrong, I refactored
this hook to no longer include a `useMemo`, by hoisting the
ReadonlyURLSearchParams call into the top-level context provider.