Fix bad decoding for x-matched-path header (#78677)
This is a follow-up to https://github.com/vercel/next.js/pull/78326
which reverts the change that interferes with the added middleware
rewrite test while also ensuring the encoding test from the
`vercel/vercel` repo is still passing as expected. The issue with the
encoding seems to be from the `x-matched-path` value being incorrectly
decoded when using non-utf8 characters so this fixes the encoding.

Fixes:
https://github.com/vercel/next.js/actions/runs/14717733032/attempts/2

<details>

<summary>test runs with patch</summary>


![CleanShot 2025-04-29 at 09 21
44@2x](https://github.com/user-attachments/assets/de5d66c5-5eb2-4e87-877e-875a7c937f35)
![CleanShot 2025-04-29 at 09 21
51@2x](https://github.com/user-attachments/assets/f1743f6c-d06c-4999-869a-9afaf1b69ab5)

</details>