Update Nissuer to alert for vulnerability reports (#78417)
### What?
We use [Nissuer](https://github.com/balazsorban44/nissuer) to assist in
triaging new GitHub Issues. This PR updates our Nissuer config to send
alerts to a webhook endpoint if Nissuer detects a possible security
report/responsible disclosure via a public GitHub Issue.

### Why?
So we can be alerted in a timely manner to possible public disclosures
of unresolved security issues in Next.js

### How?
Nissuer performs keyword detection and sends an alert to a webhook if
any keywords are detected in a newly-raised GitHub Issue.

Note: we must also set the appropriate `NISSUER_WEBHOOK_URL` and
`NISSUER_WEBHOOK_SECRET` environmental variables on the repository for
this to be effective.

Co-authored-by: JJ Kasper <jj@jjsweb.site>