Postpone on `useSearchParams` access for segment prefetch requests
Normally, when using `useSearchParams` without a suspense boundary we
throw a build error to inform a user about this mistake as early as
possible. However, when `ppr` and `clientSegmentCache` are enabled, it
can occur that a postponed page is built successfully due to accessing
fallback params, and then only the segment prefetch request (using the
actual params) fails. This results in a 500 error response, which we'd
like to avoid.

As an immediate fix to unblock further dogfooding, we can postpone when
`useSearchParams` is accessed during those on-demand static prefetch
generations.

This fix does not work when `dynamicIO` is enabled, though. For that to
work, we need to come up with a better approach.