fix: broken deploy test from #77898 (#78692)
i messed up a deployment test thing in
https://github.com/vercel/next.js/pull/77898, had the fix locally but
forgot to push before merging. the test tries to wait for some initial
stuff to happen stuff after the deployment finishes, but we should only
to do that once, it won't work in the second test of the `it.each`, so
the test fails.