Use an NFA approach for globs.
On our simple benchmarks this is 3-4x as fast as the existing glob implementation, but still 3 times slower than `fast-glob`.  My expectation is that `fast-glob` would be faster for single matches but that this should be faster when the glob is executed multiple times.  But we still need to close a gap there.