[Fix] Inverse prefetch segment for Pages routes
Another follow-up to #78387 and #78568. Refer to those PRs for
full details.

Fixes a routing conflict related to per-segment prefetches in sites
that use both App Router and Pages Router.

The fix is to expand the behavior added in the previous PRs — we output
an "inverse" tree prefetch route for every route in the route manifest
that is not PPR-enabled, including ones defined by Pages Router.