chore(CI): Split trailing-slashes suite cases into smaller suites, port to more modern e2e framework (#78890)
Same basic idea as https://github.com/vercel/next.js/pull/78787

This test suite has a lot of timeouts with rspack, which causes it to take a long time to run (though not quite long enough that it breaks the job).

In general, large test suites are good to break up, because they benefit from parallelism.

---

In addition to breaking this test suite up, I also ported it to the new framework, as the way it was patching `next.config.js` was causing problems, since the old integration tests are not run in isolated directories.