[Fix] Inverse prefetch segment for Pages routes (#78932)
Another follow-up to https://github.com/vercel/next.js/pull/78387 and
https://github.com/vercel/next.js/pull/78568. Refer to those PRs for
full details.

Fixes a routing conflict related to per-segment prefetches in sites that
use both App Router and Pages Router.

The fix is to expand the behavior added in the previous PRs — we output
an "inverse" tree prefetch route for every route in the route manifest
that is not PPR-enabled, including ones defined by Pages Router.

---------

Co-authored-by: Wyatt Johnson <accounts+github@wyattjoh.ca>