Fix tracing of server actions imported by client components (#78968)
This fixes us failing to trace actions imported by client components due
to the transform removing the actions import so our tracing isn't able
to discover the action import correctly. This was previously handled in
https://github.com/vercel/next.js/pull/73710/files#diff-f39646a8a89ce97754025c8258dc0709f67e4340c07735a3a82a87595ba01c10L526
but removed as part of the experiment.

A regression test for the `use client` importing an actions file with
file needing to be traced has been added.