chore(CI): Add Rspack labeler config (#79004)
This should automatically add the `Rspack` label in most cases, which should let us run the rspack CI in theses cases.

There's an issue with the current CI config where the build-and-test job isn't triggered by label changes, so I'll need to make some changes there, but this is a start.