chore(CI): Add back test retries (again) to arewerspackyet tests (#79014)
Test retries are important because of our high test flakiness. Without
this, we may get false-positive failures detected in the manifests and
the arewerspackyet dataset.

We had to disable test retries for arewerspackyet tests because we were
suffering from timeouts in CI. Since then we've fixed this in a number
of ways:

- I broke 5 large test suites that were contributing most to long job
execution duration when timing out into smaller test suites that can run
in parallel.
- I reduced the per-test timeout from 4 minutes back down to 1 minute
(same as the webpack tests use) for both Rspack and Turbopack.
- The Rspack team fixed the underlying regressions that were causing
increased test failures in our CI. 🙏
- Going forward, I'm working on running Rspack CI in more places, so a
regression on tests is will be much less likely in the first place.

Rspack test runs are taking <20 minutes (with a 90 minute timeout), so
it seems safe to add back the default 2 retries!

This only impacts the nightly arewerspackyet CI runs, so the CI capacity
hit from retries shouldn't be too significant.