docs: improve  `'use client'` directive and client components (#79115)
### Why?

Our docs mention the `'use client'` directive to "designate a Client Component,"  but this can easily lead to misunderstanding that the users need to mark the directive to write a Client Component. This chains to ambiguity on the concept that the entry point Client Components' props must be serializable.

It is essential to emphasize that this directive serves as the **entry point** for the client-server **boundary**, not a Client Component opt-in feature.