perf(turbopack): Optimize `next_page_static_info` (#79009)
### What?

Avoid using `FxHashMap` for the `next_page_static_info` pass and use a custom trait instead. The goal of using a custom trait is to keep it general and make the responsibility of each code.

### Why?

We do not need to allocate at all. 

Also, visitor calls are expensive, and we don't have to pay for them in this case.

### How?

 - Closes SWC-655