Reference the workUnitStore when determining bailout to CSR (#79118)
When useSearchParams is determining if it should bail out to client side
rendering it currently consults the worksStore. This is not semantically
correct since the workStore does not represent a particular render. This
change updates it to reference the workUnitStore where we can correctly
identify what type of render we are within