Always pass implicit/soft tags into the `CacheHandler.get` method
This allows a cache handler to check the timestamp of a cache entry
against the latest expiration of the given soft tags, and discarding the
entry if it's stale. Subsequently, it does not need to consult a
potentially separate tags service to retrieve a value in the
`CacheHandler.getExpiration` call for those tags, so that Next.js could
handle the expiration check. It can return `Infinity` instead.