bugfix (pages): assetPrefix should not cause hard nav in development (#79176)
In development when handling the page request, we are sending the
parsedUrl with the basePath stripped but not assetPrefix. As a result,
when setting an assetPrefix, the chunk would 404 which causes a hard
nav. It'd then recover and work correctly on subsequent navs.

This applies the exact same handling we had for basePath, but for
assetPrefix.

Fixes #77241