[turbopack] Delete the `read_glob` function (#79166)
Why? This was only used for unit tests and an example.  Also, the API was a little strange, it would end up creating a `Vc` that depended on all the directories with matches and produced a value that modeled the matching directory tree.

Happy to replace the example... I wasn't really sure what those were used for.