Repro of infinite Suspense fallback with staleTimes and middleware rewrites
I've found a scenario where a middleware rewrite that changes while the page is open can cause a infinite suspense fallback. Only happens when using `experimental.staleTImes`. Seems like it's a bug.

The PR contains a reproduction in the form form of a failing E2E test. I haven't spent time looking at your E2E testing conventions, so I may be doing something in an odd way. Also it's possible that the some parts of the code might no actually be the cause of the bug. It's an extraction from a private repo.

Hope works, otherwise I'm happy to open an issue. This just seemed like a more convenient way to demonstrate.