Use `onPostpone` to determine if segment prefetch is partial
When dynamicIO is enabled, missing data is encoded to an infinitely
hanging promise, the absence of which we use to determine if a segment
is fully static or partially static. However, when dynamicIO is not
enabled, this trick doesn't work.

Previously, if PPR is enabled, and dynamicIO is not, we were
conservative and assumed that all segments are partial. That doesn't
need to be the case, though. We can use the `onPostpone` callback of the
`prerender` function to determine if a given RSC node is partial.

To make sure that this works as expected, we're disabling `dynamicIO`
in `test/e2e/app-dir/segment-cache/incremental-opt-in`.

In addition, we're also disabling `dynamicIO` in
`test/e2e/app-dir/segment-cache/client-only-opt-in` as well as
`test/e2e/app-dir/segment-cache/export`, to prepare for an upcoming
change where `ppr` will be enabled automatically when `dynamicIO` is
enabled. Those three tests are then not compatible with `dynamicIO`
because they either rely on the `'incremental'` PPR config, or on
`output: 'export'`.