perf(turbopack): Use rayon threadpool for `minify()` (#79261)
### What?

Run `minify()` on the `rayon` threadpool instead of the `tokio` threadpool.

### Why?

ES minification is a very CPU-intensive operation, and more importantly, it calls `rayon` internally anyway. So it's better to run it from the rayon thread pool from the start.

It seems like this PR makes the total build ~ 3% faster, at the cost of a small RSS increase.

x-ref: Numbers at
https://vercel.slack.com/archives/C06PPGZ0FD3/p1747349799503749?thread_ts=1747349784.352809&cid=C06PPGZ0FD3