Client router should discard stale prefetch entries for static pages
When navigating to a static page that was previously prefetched, the
client router should discard a stale prefetch entry, and fetch the page
again, if the stale time has passed.

The selected stale time is either the default value of 5 minutes, a
custom value set via `experimental.staleTimes.static`, or a stale value
set via a `cacheLife` profile when using `"use cache"`.

fixes #74272
fixes #79093