[release] create empty changeset for next when no changeset found during canary release (#79049)
### Why?

When there are no changesets for `next`, the canary release won't be
published even if there are new commits. If this is the case, add an
empty changeset for `next` to let the canary release proceed.

Note that this will result in an empty Changelog and GitHub Release, but
this PR does not cover handling that part.

### Testing Plan

Set env `RELEASE_TYPE=canary` and run `ci:version` locally to see the
result.

---------

Co-authored-by: JJ Kasper <jj@jjsweb.site>
Co-authored-by: Sebastian "Sebbie" Silbermann <sebastian.silbermann@vercel.com>