[Segment Cache] Fix: Skew during dynamic prefetch
Fixes a bug that occurred when prefetching a Link to a dynamic page on
a different deployment. I didn't catch this earlier during dogfooding
because the app we were testing it on was using skew protection. It
also only applies to non-PPR/static routes.

This particular case happens when an app is redeployed on the server,
but it can also happen when linking to a completely different Next app.