Implement initial handler interface for pages routes (#79260)
Continuation of https://github.com/vercel/next.js/pull/78166 this
implements the `handler` interface for pages routes, This does not move
the response handling inside of the handler yet as that will be in a
follow-up PR to keep the changes isolated. This still returns the
`RenderResult` for the `base-server` to continue to handle.

Validated against `vercel/vercel` deploy tests here
https://github.com/vercel/vercel/pull/13349