[dynamicIO] Disallow sync IO in Server Components
Prior to this change sync IO had special warnings if you used it and no shell was produced but it wasn't specifically considered an error on it's own. This change makes it so that if any sync IO is detected during Server rendering you must explicitly cache it or guard against it using `await connection()`.

The rationale is that caching in RSC is such an important concept b/c it will control whether you have a static page, and how much of your page is prerenderable, that we do not want the early termination of the prerender to unecessarily degrade your prerendered experience.