[dynamicIO] Allow sync IO in client components
Prior to this change any sync IO access was considered a warning. However because React's prerender chooses to produce fallbacks before rendering children we will necessarily always have a shell before any sync IO in a client component can run if the sync IO is itself inside a Suspense boundary.

While this can still lead to spooky action at a distance where sync IO in a client component causes some other later boundary to not be prerendered as deeply as it otherwise could be it won't ever lead to an error. In the interest of being as ecosystem compatible as possible we now do not warn for sync IO specifically. sync IO in the client that does not have an ancestor Suspense boundary will still cause a regular dynamicIO validation error because you would fail to produce a shell but that is just the normal dynamicIO validation at that point and not specific to sync IO.